Refactored the send message code
Moved common messaging code to commons
