+++ packages/suibian-server/src/sockets/helper/room.ts
@@ -8,7 +8,8 @@ import {
   Position,
   VotingStatus,
 } from "@suibian/commons";
-import { sendMessage, broadcastRoom } from "./messaging";
+import { broadcastRoom } from "./messaging";
+import { sendMessage } from "@suibian/commons";
 import { foodImageQuery } from "../../queries/food";
 import { getRoomJoinQuery, joinRoomQuery } from "../../queries/join";
 import { listSocketsRoom } from "./utils";
